Fix bug of going out of bounds
This time it was caused by an edge attracting an agent that was not even in collision course
with it. The agent was facing 180 and the edge was the top wall with a small relative angle.
The solution is to only be attracted to edges that are in a potential collision course